+++ yast/lifespan.py
@@ -8,14 +8,14 @@ from yast.types import ASGIApp, Scope, Receive, Send, Message
 STATE_TRANSITION_ERROR = "Got invalid state transition on lifespan protocal"
 
 class EventType(enum.Enum):
-    STARTUP: str = 'startup'
-    CLEANUP: str = 'cleanup'
+    STARTUP:    str = 'startup'
+    SHUTDOWN:   str = 'shutdown'
+    CLEANUP:    str = 'cleanup'
 
     def __str__(self) -> str:
         return self.value
 
 class LifeSpanHandler(object):
-    
     def __init__(self) -> None:
         self.handlers = {
             et: []
@@ -32,7 +32,15 @@ class LifeSpanHandler(object):
     def add_event_handler(
             self, event_type: str, func: typing.Callable
         ) -> None:
-        self.handlers[EventType(event_type)].append(func)
+        event_types = [EventType(event_type)]
+
+        if EventType(event_type) in (EventType.CLEANUP, EventType.SHUTDOWN):
+            event_types += (EventType.CLEANUP, EventType.SHUTDOWN)
+            
+        [
+            self.handlers[et].append(func)
+            for et in set(event_types)
+        ]
     
     async def run_handler(self, event_type: EventType) -> None:
         assert event_type in self.handlers, f'EventType "{str(event_type)}" not supported'
@@ -60,10 +68,12 @@ class LifeSpanContext(object):
         ) -> None:
         self.timeout = {
             et: kwargs.get(f'{et.value}_timeout', 10) 
-            for et in list(EventType)
+            for et in list(EventType) if EventType.CLEANUP != et
         }
         self.events = {
-            et: asyncio.Event() for et in list(EventType)
+            et: asyncio.Event()
+            for et in list(EventType)
+            if EventType.CLEANUP != et
         }
 
         self.receive_queue = asyncio.Queue()
@@ -76,19 +86,20 @@ class LifeSpanContext(object):
     
     def __exit__(self, exc_type, exc, tb) -> None:
         loop = asyncio.get_event_loop()
-        loop.run_until_complete(self.wait_event_type(EventType.CLEANUP))
+        loop.run_until_complete(self.wait_event_type(EventType.SHUTDOWN))
     
     async def run_lifespan(self) -> None:
         try:
             await self.asgi(self.receive, self.send)
         finally:
-            [ events.set() for events in self.events]
+            [ event.set() for event in self.events.values()]
+                
     
     async def send(self, message: Message) -> None:
         if message['type'] == 'lifespan.startup.complete':
             self.events[EventType.STARTUP].set()
         else:
-            self.events[EventType.CLEANUP].set()
+            self.events[EventType.SHUTDOWN].set()
 
     async def receive(self) -> None:
         return await self.receive_queue.get()