+++ changelog.md
@@ -1,6 +1,7 @@
 # v0.2.3
 1. graphql: add graphiql
 2. add route: suport graphql
+3. lifespan: event_type support `shutdown`
 
 
 # v0.2.2